==> 1_city_coordinates.ipynb <==
# Installing dotenv library to manage confidential keys saved as environment variables
#!pip install python-dotenv

import requests
import json
import pandas as pd
import boto3

import os
from dotenv import load_dotenv

load_dotenv()

def get_city_name(name):
    # Extracts the name of the city from the detailed name of the destination
    return name.split(', ')[0]

# List of top-35 tourist destinations in France
best_cities_list = ["Mont Saint Michel", 
                        "St Malo",
                        "Bayeux",
                        "Le Havre",
                        "Rouen",
                        "Paris",
                        "Amiens",
                        "Lille",
                        "Strasbourg",
                        "Chateau du Haut Koenigsbourg",
                        "Colmar",
                        "Eguisheim",
                        "Besancon",
                        "Dijon",
                        "Annecy",
                        "Grenoble",
                        "Lyon",
                        "Gorges du Verdon",
                        "Bormes les Mimosas",
                        "Cassis",
                        "Marseille",
                        "Aix en Provence",
                        "Avignon",
                        "Uzes",
                        "Nimes",
                        "Aigues Mortes",
                        "Saintes Maries de la mer",
                        "Collioure",
                        "Carcassonne",
                        "Ariege",
                        "Toulouse",
                        "Montauban",
                        "Biarritz",
                        "Bayonne",
                        "La Rochelle"]


### Getting the city coordinates data

# Iterating over the list of destinations to get GPS coordinates in .json format for each destination 
# with the help of API of nominatim.openstreetmap.org

results = []

for city in best_cities_list:
    response = requests.get(f"https://nominatim.openstreetmap.org/search?q={city}&country=France&format=json")
    results.append(response.json())
    

# Checking what an element of the 'results' list looks like
# (commented to avoid long output)
#print("First element of 'results' list: ", results[0])
print()

# Checking what information about one city looks like
print("Information about one city: ", results[0][0])

# Creating a list of dictionaries (one dictionary per city)
city_list = []

for i in range(0, len(results)):
    city_info = results[i][0]
    city_list.append(city_info)

# Creating a dataframe from list of dictionaries
cities_df = pd.DataFrame.from_records(city_list)

# Checking what the resulting dataframe looks like
display(cities_df.head())


# Creating a new dataframe with only the necessary columns
city_coord = cities_df[['place_id', 'display_name', 'lat', 'lon' ]].copy()

# Checking what the resulting dataframe looks like
display(city_coord.head())


# Creating a column "city_name" that will contain only the name of the city without other details.
# We use a previously defined function "get_city_name" to do this.
city_coord['city_name'] = city_coord['display_name'].apply(get_city_name)

# Leaving only the necessary columns
city_coord=city_coord[['place_id', 'city_name','lat', 'lon']]

# Checking the resulting dataframe
display(city_coord.head())

### Saving the dataframe locally and in bucket S3

# Saving the dataframe as .csv file locally
city_coord.to_csv("city_coordinates.csv", index=False)

# (Cell is commented to avoid unnecessary rewriting in S3)
"""# Access key for user with access to write in S3 bucket
S3_ACCESS_KEY_ID =  os.getenv("S3_ACCESS_KEY_ID")
# Secret key for user with access to write in S3 bucket 
S3_SECRET_ACCESS_KEY =  os.getenv("S3_SECRET_ACCESS_KEY")

# Writing the .csv file to bucket S3
session = boto3.Session(aws_access_key_id=S3_ACCESS_KEY_ID, 
                        aws_secret_access_key=S3_SECRET_ACCESS_KEY)
s3 = session.resource("s3")
bucket = s3.Bucket("kayak-booking-bucket-12-12-2022") 
bucket.upload_file("city_coordinates.csv", Key="city_coordinates.csv")"""
