==> 2_weather_data.ipynb <==
# Installing dotenv library to manage confidential keys saved as environment variables
#!pip install python-dotenv

import requests
import json
import pandas as pd
import datetime
import boto3

import os
from dotenv import load_dotenv

load_dotenv()

def timestamp_to_datetime(df, column):
# converts 10-digit timestamp to datetime
    for row in range(0, len(df)):
        df.loc[row, column] = datetime.datetime.fromtimestamp(int(df.loc[row, column]))

def split_dict_to_columns(df, column_to_split):
    # Splits column where each cell is a dictionary 
    # into several columns (one new column per key)
    keys = df[column_to_split][0].keys()
    for row in range(0, len(df)):
        for key in keys:
            df.loc[row, column_to_split+"_"+key] = df[column_to_split][row][key]

# Loading .csv file with city coordinates from S3 bucket
url = 'https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/city_coordinates.csv'
city_coord = pd.read_csv(url)
city_coord.head(5)

### Getting weather data

results = []

# Loading API key from .env file
API_KEY = os.getenv("API_KEY")

# Using city latitude and longtitude to get weather forecast for next 7 days for each of the cities
for i in range(0, len(city_coord)):
    lat = city_coord.loc[i, "lat"]
    lon = city_coord.loc[i, "lon"]
    response = requests.get(f"https://api.openweathermap.org/data/3.0/onecall?lat={lat}&lon={lon}&exclude=current,minutely,hourly&units=metric&appid={API_KEY}")
    # We exclude current, minutely and hourly information to focus on daily forecast
    # We indicate that we want to get values expressed in metric units
    results.append(response.json())

# Checking if we obtained information for all 35 cities
print("Lenght of results: ", len(results))
print()

# Checking what information we obtained for a city (commented to avoid long output)
#print("Data obtained for a city: ", results[0])

# Checking the type of data obtained for a city
print("Type of data obtained for a city: ", type(results[0]))
print()

# As information for each one city is a dictionary, let us get a list of keys for future use
print("Dictionary keys: ", results[0].keys())
print()

# Checking information stored as value for the 'daily' key
# (commented to avoid long output)
#print("Value stored in 'daily' key: ", results[0]['daily'])
print()

# The information is stored as a list
print("Data type of value stored in 'daily' key: ", type(results[0]['daily']))
print()

# The length of the list corresponds to the number of days for which the forecast is know
print("Number of days: ",len(results[0]['daily']))
print()

# Weather data for one day in one city
print("Forecast for one day: ", results[0]['daily'][0])
print()

print("Data type for one day of weather data: ", type(results[0]['daily'][0]))

Weather data for each day is stored in the form of dictionary. We will create a list of such dictionaries and then build a dataframe from this list.


# Creating a list of dictionaries
weather_by_day = []

for city in range(0, len(results)):
    for day in range(0, 7): 
        # keeping weather information about next 7 days
        day_weather = results[city]['daily'][day]
        weather_by_day.append(day_weather)

# Creating a dataframe from a list of dictionaries
weekly_forecast_df = pd.DataFrame.from_records(weather_by_day)

# Checking the resulting dataframe
weekly_forecast_df.head()

# Checking what columns contain data in the string ('object') form
weekly_forecast_df.select_dtypes(include=['object']).head()

# Checking the contents of a cell in 'temp' column
print(weekly_forecast_df.loc[0, "temp"])
(print)

# Checking the contents of a cell in 'feels_like' column
print(weekly_forecast_df.loc[0, "feels_like"])

# Checking the contents of a cell in 'weather' column
print(weekly_forecast_df.loc[0, "weather"])

Each of the cells in these columns contains a dictionary. I would like to split this information into several columns, so that each key in such dictionaries becomes a separate column.

# Splitting columns using the previously defined function 'split_dict_to_columns'
split_dict_to_columns(weekly_forecast_df, "temp")

split_dict_to_columns(weekly_forecast_df, "feels_like")

In the 'weather' column, the cell contains a list that contains a dictionary, so there is a slight adjustment to the code compared to other columns.

# Keys for creating columns from splitting 'weather' column
weather_keys = weekly_forecast_df["weather"][0][0].keys()
print(weather_keys)

# Splitting dictionary in column 'weather' into several columns
for row in range(0, len(weekly_forecast_df)): 
    for key in weather_keys:
        weekly_forecast_df.loc[row, "weather_"+key] = weekly_forecast_df["weather"][row][0][key]


# Checking the resulting table (commented to avoid long output)
#pd.set_option('display.max_columns', None)
#weekly_forecast_df.head(5)

print(weekly_forecast_df.columns)

# Dropping some unnecessary columns
columns_to_drop = ['moonrise', 'moonset', 'moon_phase', 'temp',
       'feels_like', 'weather', 'weather_icon']

weather_df = weekly_forecast_df.drop(columns_to_drop, axis=1)

Some columns contain information about time in timestamp format. We will convert this information into datetime format.

timestamp_date_columns = ['dt', 'sunrise', 'sunset']
for column in timestamp_date_columns:
    timestamp_to_datetime(weather_df, column)

'weather_df' dataframe contains 245 rows (35 cities x 7 days). To be able to merge the table easily with the other tables in this project, we will create a column 'city_name' for each of the rows in the 'weather_df'. The weather data was obtained based on the list of cities in the table 'city_coord', so we can simply create a column using the same list with each city name repeated 7 times.

print("Length of the 'weather_df': ", len(weather_df))

# Adding column with city name for each row with daily forecast
target_df_start_index = 0
target_df_end_index = 7
for source_df_row in range(0, len(city_coord)):
    for target_df_row in range(target_df_start_index, target_df_end_index):
        weather_df.loc[target_df_row, "city_name"] = city_coord.loc[source_df_row, "city_name"]
    target_df_start_index = target_df_start_index+7
    target_df_end_index= target_df_end_index+7


# Checking the resulting dataframe (commented to avoid long output)
# weather_df.head(10)

### Saving the dataframe locally and in bucket S3

# Saving the dataframe as .csv file locally
weather_df.to_csv("weather_forecast.csv", index=False)

# (Cell is commented to avoid unnecessary rewriting in S3)
"""# Access key for user with access to write in S3 bucket
S3_ACCESS_KEY_ID =  os.getenv("S3_ACCESS_KEY_ID")
# Secret key for user with access to write in S3 bucket 
S3_SECRET_ACCESS_KEY =  os.getenv("S3_SECRET_ACCESS_KEY")

# Writing the .csv file to bucket S3
session = boto3.Session(aws_access_key_id=S3_ACCESS_KEY_ID, 
                      aws_secret_access_key=S3_SECRET_ACCESS_KEY)
s3 = session.resource("s3")
bucket = s3.Bucket("kayak-booking-bucket-12-12-2022") 
bucket.upload_file("weather_forecast.csv", Key="weather_forecast.csv")"""
