==> 4_kayak_df.ipynb <==
# Installing dotenv library to manage confidential keys saved as environment variables
#!pip install python-dotenv

import json
import pandas as pd
import re
import datetime as dt

import boto3

import os
from urllib.request import urlopen
from dotenv import load_dotenv

load_dotenv()

### Creating a table with information about hotels

# Loading .json file with hotel urls
url = "https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/booking_search_page.json"
response_hotel_urls = urlopen(url)
hotel_urls = json.loads(response_hotel_urls.read())

# Creating a dataframe from .json file
hotel_urls_df = pd.DataFrame.from_records(hotel_urls)

display(hotel_urls_df.head())

# Checking the length of the dataframe
print("Length of the dataframe:", len(hotel_urls_df))

The "booking_city_url" column contains the url that was used to obtain a list of hotels in a given city. In order to know the city where each of the hotels is situated, let us extract the name of the city from the url.

# Example of text in column "booking_city_url"
print(hotel_urls_df.iloc[0]["booking_city_url"])

# Splitting the url using "=" and "&" as delimiters and conserving part of url as the name of the city
hotel_urls_df["city_name"] = hotel_urls_df["booking_city_url"].apply(lambda x: re.split(r"=|&", x)[1])

# Replacing "+" with " " in the names of cities consisting of several words
hotel_urls_df["city_name"] = hotel_urls_df["city_name"].apply(lambda x: x.replace("+", " "))

# Dropping "booking_city_url" column
hotel_urls_df = hotel_urls_df.drop("booking_city_url", axis=1)

hotel_urls_df.head()

We will now use join the table containing hotel urls with the table containing other data for each hotel (score, address, description etc).

# Loading .json file with hotels data
url = "https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/booking_hotels_data.json"
response_hotels_data = urlopen(url)
hotels = json.loads(response_hotels_data.read())

# Creating a dataframe from .json file
hotels_df = pd.DataFrame.from_records(hotels)

hotels_df.head()

# Renaming columns for clarity
hotels_df.rename(columns = {'latitide':'hotel_latitude'}, inplace = True)
hotels_df.rename(columns = {'longtitude':'hotel_longtitude'}, inplace = True)

# Addresses of hotels contain html tag /n, we will use .strip() in  order to get rid of it
hotels_df["location"] = hotels_df["location"].apply(lambda x: str.strip(x))

# Checking the resulting dataframe
hotels_df.head()

Some hotel descriptions begin with promotional information about a discount. We will remove these phrases from descriptions.

# Example of hotel description with promotional information
print("Description before cleaning:")
print("'", hotels_df.iloc[3]["description"], "'")
print()

# If the description contains words "Genius discount", we only keep the text the follows the phrase ""To save at this property, all you have to do is  ."
hotels_df["description"] = hotels_df["description"].apply(lambda x: x.split("To save at this property, all you have to do is  . ")[1] if 'Genius discount' in x else x)

print("Description after cleaning: ")
print("'", hotels_df.iloc[3]["description"], "'")

# Removing part of text from "hotel_url" column and keeping only the cleaned url
hotels_df["hotel_url"] = hotels_df["hotel_url"].apply(lambda x: x.strip("<HtmlResponse 200 ").split("?")[0])

# Some hotel urls contain "en_gb." part indicating that the url corresponds to the English-language version of the page. 
print("Example of url:", hotels_df.iloc[5]["hotel_url"])

# The hotel url remains functional without the part that indicates the language version of the page,
# so we'll remove this part of string for uniformity (the url remains functional).
hotels_df["hotel_url"] = hotels_df["hotel_url"].apply(lambda x: x.replace("en-gb.", "") if "en-gb." in x else x)

hotels_df.head()

Merging the table containing hotel urls and the table containing other information about hotels

# Rearranging order of columns to prepare the table for the merge with hotel_urls_df
hotels_df = hotels_df[['hotel_url', 'hotel_name', 'score', 'description', 'location', 'hotel_latitude',
       'hotel_longtitude']]

# Merging tables
hotel_info_df = hotel_urls_df.merge(hotels_df, how="left")

# Checking the resulting table
hotel_info_df.head()

# Checking if there are any missing values in the table
hotel_info_df.isnull().sum()

It turns out that there are a number of hotels with missing scores. Let's have a closer look at them to see if there's a problem.

display(hotel_info_df[hotel_info_df["score"].isnull()].head())

# Saving indexes of rows where the hotel score is missing
no_score_indexes = hotel_info_df[hotel_info_df["score"].isnull()].index

If we print the descriptions of hotels that have no score, it turns out that most of these are in fact apartments or villas.
If they have been present on booking.com not for a long time, it is normal that they do not have a score yet as there are many less visitors than in hotels.

# (code below commented to avoid long output)
"""for index in no_score_indexes:
    print(hotel_info_df["description"].iloc[index])
    print()"""

### Enriching the table with city coordinates

# Loading .csv file with city coordinates from S3 bucket
url = 'https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/city_coordinates.csv'
city_coord = pd.read_csv(url)
city_coord.head(5)

The names of the city in the table "city_coord" are written with accents, hyphens and use "Saint" instead of "St" in names like "Saint-Malo". I would like to keep this variant in the final table.
I will replace the names of the cities in "hotel_info_df" table with the names of the cities as they are written in "city_coord" table. Then I'll use the city names as keys to merge the two tables.

# Saving city names as a list
city_names_coord= city_coord["city_name"].tolist()

# Sorting city names list in the alphabetical order
city_names_coord.sort()

# We replace St with Saint (otherwise, the cities in the two lists will not be sorted in identical order)
hotel_info_df["city_name"] = hotel_info_df["city_name"].apply(lambda x: x.replace("St ", "Saint ") if "St " in x else x)

# Saving city names from "hotel_iinfo_df" table as a list
city_names_hotel = hotel_info_df["city_name"].unique().tolist()

# Sorting city names list in the alphabetical order
city_names_hotel.sort()

# Checking if the order of cities in the two lists is correct
print(list(zip(city_names_hotel, city_names_coord)))

# "Marseille" is spelled as "Marseilla" in one of the lists, so we'll correct the mistake
city_names_coord = [city.replace("Marseilla", "Marseille") for city in city_names_coord]

# Replacing city names in "hotel_info_df" table
for i in range(0, len(hotel_info_df["city_name"])):
    if hotel_info_df["city_name"][i] in city_names_hotel:
        index = city_names_hotel.index(hotel_info_df["city_name"].loc[i])
        hotel_info_df["city_name"] = hotel_info_df["city_name"].replace(hotel_info_df["city_name"][i], city_names_coord[index])
    else: 
        pass
        

# Checking the resulting table
hotel_info_df.head()

# Checking the order of columns 
print(city_coord.columns)

# Rearranging the order of columns to prepare for merging with "hotel_info_df"
city_coord = city_coord[['place_id',  'lat', 'lon', 'city_name']]

# Merging the two tables
city_and_hotel_info_df = city_coord.merge(hotel_info_df, how="right")

# Checking the order of the columns in the resulting table
print(city_and_hotel_info_df.columns)

# Rearranging the order of the columns for clarity
city_and_hotel_info_df = city_and_hotel_info_df[['city_name', 'place_id', 'lat', 'lon', 'hotel_name', 'hotel_url',
       'score', 'description', 'location', 'hotel_latitude', 'hotel_longtitude']]

# Checking the resulting table
city_and_hotel_info_df.head()

### Enriching the table with weather data

We need to add 7-day weather forecast for each hotel based on the city where the hotel is situated. To do this, we repeat each line containing information about a hotel 7 times.

# Saving the number of hotels in a variable, we will need it later
nb_hotels = len(city_and_hotel_info_df)

# Repeating every line 7 times
city_and_hotel_info_df = city_and_hotel_info_df.loc[city_and_hotel_info_df.index.repeat(7)]

# Checking the result
city_and_hotel_info_df.head(10)

We wil now create a key that will later help us merge the dataframe with the hotel information and the dataframe with the weather information.

# Creating a list of seven numbers that correspont to the next seven days
days = [1, 2, 3, 4, 5, 6, 7]

# Creating a column that will contain, for each hotel, a number of the day
day_numbers = days * nb_hotels
city_and_hotel_info_df["day_in_city"] = day_numbers

# For each day, we add the name of the city where the hotel is situated. 
# This column will be used to indicate weather forecast for each hotel based on the city.
city_and_hotel_info_df["day_in_city"] = city_and_hotel_info_df['city_name'] + " day "+ city_and_hotel_info_df["day_in_city"].astype(str)

# Checking the resulting table
city_and_hotel_info_df.head()

# Loading .csv file with weather data from S3 bucket
url = 'https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/weather_forecast.csv'
weather_df = pd.read_csv(url)

# Each line of the table is contains weather information for a specific city
# and for a specific date for the next 7 days.
weather_df.head(10)

# We create the columnn which the city name and the number of the day
nb_cities = weather_df["city_name"].nunique()
days = [1, 2, 3, 4, 5, 6, 7]
day_numbers = days * nb_cities
weather_df["day_in_city"] = day_numbers
weather_df["day_in_city"] = weather_df['city_name'] + " day "+ weather_df["day_in_city"].astype(str)

# Checking the result
weather_df.tail(10)

# Checking the order of columns in the dataframe
print(weather_df.columns)

# Changing the order of columns in order to be able to merge dataframes easily.
# We will keep 'city_name' column in both dataframes in order to check later if the merge has been done correctly.
weather_df = weather_df[['day_in_city', 'city_name', 'dt', 'sunrise', 'sunset', 'pressure', 'humidity',
       'dew_point', 'wind_speed', 'wind_deg', 'wind_gust', 'clouds', 'pop',
       'uvi', 'rain', 'snow', 'temp_day', 'temp_min', 'temp_max', 'temp_night',
       'temp_eve', 'temp_morn', 'feels_like_day', 'feels_like_night',
       'feels_like_eve', 'feels_like_morn', 'weather_id', 'weather_main',
       'weather_description']]

# Merging the dataframes
kayak_df = city_and_hotel_info_df.merge(weather_df, how="left")

# Checking the result (commented to avoid long output)
#pd.set_option('display.max_columns', None)
#kayak_df.tail(15)

Now we have the dataframe with all the necessary information, but some adjustments have to be made.

# Renaming some columns for clarity
kayak_df.rename(columns = {'place_id':'city_id'}, inplace = True)
kayak_df.rename(columns = {'lat':'city_latitude'}, inplace = True)
kayak_df.rename(columns = {'lon':'city_longtitude'}, inplace = True)
kayak_df.rename(columns = {'dt':'date'}, inplace = True)

Values in some columns are still in string format while they should be in float or datetime format:

# Checking data types present in the dataframe (commented to avoid long output)
#kayak_df.dtypes

# Converting the values in the columns "date", "sunrise", "sunset" to datetime type
columns_to_convert = ["date", "sunrise", "sunset"]
for column in columns_to_convert: 
    kayak_df[column] = pd.to_datetime(kayak_df[column], format='%Y-%m-%d %H:%M:%S')

# Creating a column with the name of the day of the week    
kayak_df["day_of_week"] = kayak_df["date"].apply(lambda x: x.day_name())

# Checking what are the remaining columns that contain values in string ("object") format
print((kayak_df.select_dtypes(include=['object'])).columns)

# Converting the values in the columns 'score', 'hotel_latitude', 'hotel_longtitude' to float type
columns_to_convert = ['score', 'hotel_latitude', 'hotel_longtitude']
for column in columns_to_convert: 
    kayak_df[column] = kayak_df[column].astype(float)

# Checking data types (commented to avoid long output)
#kayak_df.dtypes

# Checking the names of columns
print(kayak_df.columns)

# Rearranging the order of columns for clarity
kayak_df = kayak_df[['city_name', 'city_id', 'city_latitude', 'city_longtitude',
       'hotel_name', 'hotel_url', 'score', 'description', 'location',
       'hotel_latitude', 'hotel_longtitude', 'date', 'day_of_week','day_in_city','sunrise',
       'sunset', 'weather_main', 'weather_description', 'weather_id','pressure', 
       'humidity', 'dew_point', 'wind_speed', 'wind_deg',
       'wind_gust', 'clouds', 'pop', 'uvi', 'rain', 'snow', 'temp_day',
       'temp_min', 'temp_max', 'temp_night', 'temp_eve', 'temp_morn',
       'feels_like_day', 'feels_like_night', 'feels_like_eve','feels_like_morn']]

Saving the .csv file locally in the working folder

kayak_df.to_csv("kayak.csv", index=False)

Saving the .csv file to S3 bucket

# (The cell is commented to avoid unnecessary rewriting).
"""# Access key for user with access to write in S3 bucket
S3_ACCESS_KEY_ID =  os.getenv("S3_ACCESS_KEY_ID")
# Secret key for user with access to write in S3 bucket 
S3_SECRET_ACCESS_KEY =  os.getenv("S3_SECRET_ACCESS_KEY")

# Writing the .csv file to bucket S3
session = boto3.Session(aws_access_key_id=S3_ACCESS_KEY_ID, 
                      aws_secret_access_key=S3_SECRET_ACCESS_KEY)
s3 = session.resource("s3")
bucket = s3.Bucket("kayak-booking-bucket-12-12-2022") 
bucket.upload_file("kayak.csv", Key="kayak.csv")"""
