==> 5_creating_sql_database_in_rds.ipynb <==
# Installing dotenv library to manage confidential keys saved as environment variables
#!pip install python-dotenv

from sqlalchemy import create_engine
import pandas as pd
from sqlalchemy.sql import text

import os
from dotenv import load_dotenv

load_dotenv()

### Creating a SQLAlchemy engine

RDS_LOGIN = os.getenv("RDS_LOGIN")
RDS_PASSWORD = os.getenv("RDS_PASSWORD")
RDS_DATABASE_NAME = os.getenv("RDS_DATABASE_NAME")

# Connecting to a remote Postrgesql instance on Amazon RDS
engine = create_engine(f"postgresql+psycopg2://{RDS_LOGIN}:{RDS_PASSWORD}@{RDS_DATABASE_NAME}.crwzncqmo9mf.eu-west-3.rds.amazonaws.com/postgres", echo=False)

# Reading .csv file from S3 bucket and storing the data in a DataFrame
url = "https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/kayak.csv"
dataset = pd.read_csv(url)

# Writing records stored in a DataFrame to a SQL database
dataset.to_sql("kayak", engine, if_exists="replace")

### Examples of SQL queries

# Connecting to the SQL database
conn = engine.connect()

# Displaying the list of destinations present in the database
statement = text("SELECT DISTINCT city_name FROM kayak")
result = conn.execute(statement)
result.fetchall()

# Displaying 10 hotels that have scores higher than 9.5
statement = text("SELECT DISTINCT hotel_name, score FROM kayak WHERE score > 9.5 LIMIT 10")
result = conn.execute(statement)
result.fetchall()
