==> 6_maps.ipynb <==
import pandas as pd
import plotly.express as px


# Loading data from S3 bucket
url = 'https://kayak-booking-bucket-12-12-2022.s3.eu-west-3.amazonaws.com/kayak.csv'
kayak_df = pd.read_csv(url)

For each destination, the table contains daily forecasts for next 7 days for each city. Let us see group toghether the forecast by the city to obtain mean values of different weather characteristics for the next 7 days.

weekly_forecast_by_city = kayak_df.groupby(['city_name']).mean(numeric_only=True).reset_index()
weekly_forecast_by_city.head()

I think that the perfect weather for visiting a city is about 25°C. So I will choose five cities where mean temperature for next 7 days is closest to 25°C.

weekly_forecast_by_city["dif_from_optimum"] = abs(25 - weekly_forecast_by_city["temp_day"])
best_cities_df = weekly_forecast_by_city[["city_name", "temp_day", "dif_from_optimum"]].sort_values(by=['dif_from_optimum'])[:5]
display(best_cities_df)

best_cities = best_cities_df["city_name"].tolist()
best_cities

weekly_forecast_by_city.loc[weekly_forecast_by_city['city_name'].isin(best_cities)]

Creating a map that will show where the cities with the best weather are situated.

fig = px.scatter_mapbox(weekly_forecast_by_city.loc[weekly_forecast_by_city['city_name'].isin(best_cities)], 
                        lat="city_latitude", lon="city_longtitude", color_continuous_scale=px.colors.sequential.Bluered_r,
                        hover_name="city_name", size = "temp_day", color="dif_from_optimum", zoom=4, mapbox_style="open-street-map")
fig.show()

Among the hotels situated in 5 best cities, let us choose 20 hotels with the best scores.

best_hotels_df = kayak_df[kayak_df['city_name'].isin(best_cities)].groupby(['hotel_name']).mean(numeric_only=True).sort_values(by=['score'], ascending=False).reset_index()[:20]
best_hotels_df.head()

best_hotels = best_hotels_df["hotel_name"].tolist()

# Displaying names of 5 hotels with the best score
best_hotels[:5]

Creating a map that shows top-20 hotels in the area

fig = px.scatter_mapbox(kayak_df.loc[kayak_df['hotel_name'].isin(best_hotels)], 
                        lat="hotel_latitude", lon="hotel_longtitude", color_continuous_scale=px.colors.sequential.Bluered,
                        hover_name="hotel_name", color="score", size="score",
                        zoom=4, mapbox_style="open-street-map")
fig.show()
